move raw entities and repos into dedicated package
