+++ bootstrap/src/test/java/com/onlydust/marketplace/indexer/bootstrap/it/IssueIndexingIT.java
@@ -4,14 +4,14 @@ import com.onlydust.marketplace.indexer.domain.models.raw.RawIssue;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawRepo;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawSocialAccount;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawUser;
-import com.onlydust.marketplace.indexer.postgres.entities.Issue;
-import com.onlydust.marketplace.indexer.postgres.entities.Repo;
-import com.onlydust.marketplace.indexer.postgres.entities.User;
-import com.onlydust.marketplace.indexer.postgres.entities.UserSocialAccounts;
-import com.onlydust.marketplace.indexer.postgres.repositories.IssueRepository;
-import com.onlydust.marketplace.indexer.postgres.repositories.RepoRepository;
-import com.onlydust.marketplace.indexer.postgres.repositories.UserRepository;
-import com.onlydust.marketplace.indexer.postgres.repositories.UserSocialAccountsRepository;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.Issue;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.Repo;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.User;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.UserSocialAccounts;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.IssueRepository;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.RepoRepository;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.UserRepository;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.UserSocialAccountsRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.reactive.server.WebTestClient;
