+++ bootstrap/src/test/java/com/onlydust/marketplace/indexer/bootstrap/it/PullRequestIndexingIT.java
@@ -1,8 +1,8 @@
 package com.onlydust.marketplace.indexer.bootstrap.it;
 
 import com.onlydust.marketplace.indexer.domain.models.raw.*;
-import com.onlydust.marketplace.indexer.postgres.entities.*;
-import com.onlydust.marketplace.indexer.postgres.repositories.*;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.*;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.*;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.reactive.server.WebTestClient;
