+++ bootstrap/src/test/java/com/onlydust/marketplace/indexer/bootstrap/it/UserIndexingIT.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawSocialAccount;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawUser;
-import com.onlydust.marketplace.indexer.postgres.entities.User;
-import com.onlydust.marketplace.indexer.postgres.entities.UserSocialAccounts;
-import com.onlydust.marketplace.indexer.postgres.repositories.UserRepository;
-import com.onlydust.marketplace.indexer.postgres.repositories.UserSocialAccountsRepository;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.User;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.UserSocialAccounts;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.UserRepository;
+import com.onlydust.marketplace.indexer.postgres.repositories.raw.UserSocialAccountsRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
