+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/entities/raw/PullRequestCommits.java
@@ -1,11 +1,8 @@
-package com.onlydust.marketplace.indexer.postgres.entities;
+package com.onlydust.marketplace.indexer.postgres.entities.raw;
 
-import com.onlydust.marketplace.indexer.domain.models.raw.RawLanguages;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawCommit;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.Type;
 import org.hibernate.annotations.TypeDef;
@@ -16,6 +13,7 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.time.Instant;
+import java.util.List;
 
 
 @Data
@@ -23,24 +21,27 @@ import java.time.Instant;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "repo_languages", schema = "indexer_raw")
+@EqualsAndHashCode
+@Table(name = "pull_request_commits", schema = "indexer_raw")
 @TypeDef(name = "jsonb", typeClass = JsonBinaryType.class)
-public class RepoLanguages {
+public class PullRequestCommits {
     @Id
-    Long repoId;
+    Long pullRequestId;
 
     @Type(type = "jsonb")
-    RawLanguages data;
+    List<RawCommit> data;
 
+    @EqualsAndHashCode.Exclude
     @CreationTimestamp
     @Column(name = "created_at", nullable = false, updatable = false)
     Instant createdAt;
 
+    @EqualsAndHashCode.Exclude
     @UpdateTimestamp
     @Column(name = "updated_at", nullable = false)
     Instant updatedAt;
 
-    public static RepoLanguages of(Long repoId, RawLanguages languages) {
-        return RepoLanguages.builder().repoId(repoId).data(languages).build();
+    public static PullRequestCommits of(Long pullRequestId, List<RawCommit> commits) {
+        return PullRequestCommits.builder().pullRequestId(pullRequestId).data(commits).build();
     }
 }
