+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/entities/raw/PullRequestReview.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.entities;
+package com.onlydust.marketplace.indexer.postgres.entities.raw;
 
-import com.onlydust.marketplace.indexer.domain.models.raw.RawIssue;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawCodeReview;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
@@ -13,26 +13,23 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.time.Instant;
+import java.util.List;
 
 
 @Data
 @Entity
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @EqualsAndHashCode
-@Table(name = "issues", schema = "indexer_raw")
+@Table(name = "pull_request_reviews", schema = "indexer_raw")
 @TypeDef(name = "jsonb", typeClass = JsonBinaryType.class)
-public class Issue {
+public class PullRequestReview {
     @Id
-    Long id;
-
-    Long repoId;
-
-    Long number;
+    Long pullRequestId;
 
     @Type(type = "jsonb")
-    RawIssue data;
+    List<RawCodeReview> data;
 
     @EqualsAndHashCode.Exclude
     @CreationTimestamp
@@ -44,7 +41,7 @@ public class Issue {
     @Column(name = "updated_at", nullable = false)
     Instant updatedAt;
 
-    public static Issue of(Long repoId, RawIssue issue) {
-        return Issue.builder().id(issue.getId()).repoId(repoId).number(issue.getNumber()).data(issue).build();
+    public static PullRequestReview of(Long pullRequestId, List<RawCodeReview> codeReviews) {
+        return PullRequestReview.builder().pullRequestId(pullRequestId).data(codeReviews).build();
     }
 }
