+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/entities/raw/Repo.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.entities;
+package com.onlydust.marketplace.indexer.postgres.entities.raw;
 
-import com.onlydust.marketplace.indexer.domain.models.raw.RawCommit;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawRepo;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
@@ -13,7 +13,6 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.time.Instant;
-import java.util.List;
 
 
 @Data
@@ -22,14 +21,18 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-@Table(name = "pull_request_commits", schema = "indexer_raw")
+@Table(name = "repos", schema = "indexer_raw")
 @TypeDef(name = "jsonb", typeClass = JsonBinaryType.class)
-public class PullRequestCommits {
+public class Repo {
     @Id
-    Long pullRequestId;
+    Long id;
+
+    String owner;
+
+    String name;
 
     @Type(type = "jsonb")
-    List<RawCommit> data;
+    RawRepo data;
 
     @EqualsAndHashCode.Exclude
     @CreationTimestamp
@@ -41,7 +44,7 @@ public class PullRequestCommits {
     @Column(name = "updated_at", nullable = false)
     Instant updatedAt;
 
-    public static PullRequestCommits of(Long pullRequestId, List<RawCommit> commits) {
-        return PullRequestCommits.builder().pullRequestId(pullRequestId).data(commits).build();
+    public static Repo of(RawRepo repo) {
+        return Repo.builder().id(repo.getId()).owner(repo.getOwner().getLogin()).name(repo.getName()).data(repo).build();
     }
 }
