--- infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/UserRepository.java
@@ -1,8 +1,0 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
-
-import com.onlydust.marketplace.indexer.postgres.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
