+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/InstallationEventRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.InstallationEvent;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.InstallationEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstallationEventRepository extends JpaRepository<InstallationEvent, Long> {
