+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/PullRequestClosingIssueRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.PullRequestClosingIssue;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.PullRequestClosingIssue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
