+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/PullRequestCommitsRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.PullRequestCommits;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.PullRequestCommits;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PullRequestCommitsRepository extends JpaRepository<PullRequestCommits, Long> {
