+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/PullRequestReviewsRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.PullRequestReview;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.PullRequestReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PullRequestReviewsRepository extends JpaRepository<PullRequestReview, Long> {
