+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/RepoCheckRunsRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.RepoCheckRuns;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.RepoCheckRuns;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RepoCheckRunsRepository extends JpaRepository<RepoCheckRuns, RepoCheckRuns.Id> {
