+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/RepoLanguagesRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.RepoLanguages;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.RepoLanguages;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RepoLanguagesRepository extends JpaRepository<RepoLanguages, Long> {
