+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/UserRepository.java
@@ -1,0 +1,8 @@
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
+
+import com.onlydust.marketplace.indexer.postgres.entities.raw.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
