+++ infrastructure/postgres/src/main/java/com/onlydust/marketplace/indexer/postgres/repositories/raw/UserSocialAccountsRepository.java
@@ -1,6 +1,6 @@
-package com.onlydust.marketplace.indexer.postgres.repositories;
+package com.onlydust.marketplace.indexer.postgres.repositories.raw;
 
-import com.onlydust.marketplace.indexer.postgres.entities.UserSocialAccounts;
+import com.onlydust.marketplace.indexer.postgres.entities.raw.UserSocialAccounts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserSocialAccountsRepository extends JpaRepository<UserSocialAccounts, Long> {
